+++ bindings/angular2/CHANGELOG.md
@@ -12,6 +12,7 @@ dev
 
 ### Misc
 
+ * Added `@angular/forms` as a new dependency.
  * OnsInput: `[(value)]` now listens to `input` event instead of `change` event.
 
 4.0.0-rc.1